Move cri/client.py to cri.py
