+++ granulate_utils/containers/client.py
@@ -6,7 +6,7 @@
 from typing import List
 
 from granulate_utils.containers.container import Container
-from granulate_utils.containers.cri.client import CriClient
+from granulate_utils.containers.cri import CriClient
 from granulate_utils.containers.docker import DockerClient
 from granulate_utils.exceptions import NoContainerRuntimesError
